+++ src/lib.rs
@@ -81,8 +81,9 @@ pub const CODEC_ID_STR: &str = "tga";
 
 pub use decoder::{
     parse_tga, parse_tga_attributes_type, parse_tga_colour_correction_table,
-    parse_tga_developer_area, parse_tga_extension_area, parse_tga_footer, parse_tga_image_id,
-    parse_tga_postage_stamp, parse_tga_scan_line_table, resolve_alpha_with_targa32_fallback,
+    parse_tga_developer_area, parse_tga_extension_area, parse_tga_footer, parse_tga_gamma,
+    parse_tga_image_id, parse_tga_key_color, parse_tga_pixel_aspect_ratio, parse_tga_postage_stamp,
+    parse_tga_scan_line_table, parse_tga_software_version, resolve_alpha_with_targa32_fallback,
 };
 pub use encoder::{
     encode_tga_grayscale, encode_tga_grayscale_rle, encode_tga_palette, encode_tga_palette_rle,
@@ -93,11 +94,11 @@ pub use encoder::{
 pub use error::{Result, TgaError};
 pub use image::{TgaImage, TgaPixelFormat};
 pub use types::{
-    parse_extension_area, parse_footer, parse_header, AttributesType, ImageType,
-    TgaColourCorrectionTable, TgaDeveloperArea, TgaDeveloperTag, TgaExtensionArea, TgaFooter,
-    TgaHeader, TgaScanLineTable, TgaTimestamp, TGA_COLOUR_CORRECTION_TABLE_ENTRIES,
-    TGA_COLOUR_CORRECTION_TABLE_SIZE, TGA_EXTENSION_AREA_SIZE, TGA_FOOTER_MAGIC, TGA_FOOTER_SIZE,
-    TGA_HEADER_SIZE,
+    parse_extension_area, parse_footer, parse_header, AttributesType, GammaValue, ImageType,
+    KeyColor, PixelAspectRatio, SoftwareVersion, TgaColourCorrectionTable, TgaDeveloperArea,
+    TgaDeveloperTag, TgaExtensionArea, TgaFooter, TgaHeader, TgaScanLineTable, TgaTimestamp,
+    TGA_COLOUR_CORRECTION_TABLE_ENTRIES, TGA_COLOUR_CORRECTION_TABLE_SIZE, TGA_EXTENSION_AREA_SIZE,
+    TGA_FOOTER_MAGIC, TGA_FOOTER_SIZE, TGA_HEADER_SIZE,
 };
 
 #[cfg(feature = "registry")]